requirements.txt
# UIDE Forense AI - Backend Dependencies
# Flask + React Migration

# ==========================================
# Flask Framework
# ==========================================
Flask>=3.0.0
flask-cors>=4.0.0

# ==========================================
# Deep Learning - PyTorch
# ==========================================
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
timm>=0.9.0

# ==========================================
# HuggingFace - Modelos Pre-entrenados
# ==========================================
transformers>=4.30.0

# ==========================================
# Procesamiento de Audio
# ==========================================
librosa>=0.10.0
scipy>=1.10.0

# ==========================================
# Visión por Computadora
# ==========================================
opencv-python>=4.8.0
pillow>=10.0.0
piexif>=1.1.3

# ==========================================
# Utilidades Generales
# ==========================================
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.31.0
matplotlib>=3.7.0

# ==========================================
# DeepSeek LLM Integration
# ==========================================
aiohttp>=3.9.0
tenacity>=8.2.0

# ==========================================
# Production Server
# ==========================================
gunicorn>=21.2.0
python-dotenv>=1.0.0
